Move str_to_size_t back to impl and address feedback on strtoul
